+++ test/unit/config.js
@@ -1,6 +1,7 @@
 const {
     detect
 } = require('detect-browser');
+import * as path from 'path';
 import parallel from 'mocha.parallel';
 
 let useParallel = describe;
@@ -19,6 +20,8 @@ const config = {
 if (config.platform.name === 'node') {
     process.setMaxListeners(100);
     require('events').EventEmitter.defaultMaxListeners = 100;
+    config.rootPath = path.join(__dirname, '../../');
+    console.log('rootPath: ' + config.rootPath);
 }
 
 export default config;